+++ CHANGES.md
@@ -6,6 +6,8 @@
 
 - Correct max string length calculation when there are string operators (#2292)
 - Fixed option usage when using the `--code` flag (#2259)
+- Do not call `uvloop.install()` when _Black_ is used as a library (#2303)
+- Added `--required-version` option to require a specific version to be running (#2300)
 - Configuration files using TOML features higher than spec v0.5.0 are now supported
   (#2301)